services/fruxAI/api/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pdfplumber>=0.10.0
python-magic>=0.4.27
aiofiles>=23.2.1
prometheus-client>=0.19.0
structlog>=23.2.0
tenacity>=8.2.0
celery>=5.3.0
